Move _ParameterSelectorMixin to components/kernel.py
